First test run is successful -- The search is now narrowed to just what is selected in the attributes list.
This needs to be refined a bit because the first time the attribute list is created, ALL attributes are selected. The way I get a bounding box is via a WPS call to GeoServer with a filter on each selected attribute. This takes a while when for example it's 50 states.  In the future I will probably get the initial bbox from the WFS getcapabilities since it will be the same as select * features.
Issue #GDP-453 - Add bbox of selected feature/s to CSW search to narrow returned data sets.

git-svn-id: https://cida-svn.er.usgs.gov/repos/dev/usgs/gdp/trunk@19919 622dc1fb-2224-0410-ab72-dc85a3975e1a