Further refined the BBOX constraint processing.  Tested to ensure that a BBOX gets created for draw layers as well as uploaded shapefiles. We are now initially grabbing the bounding box for the layer when a user chooses a new layer from the GetCapabilities that's cached when the application loads. If a user selects an attibute or more than one attribute, a WPS request is sent to GeoServer for the bounding box covering the combined selected attributes.
This boundingbox is used as a constraint to the CSW search.  The boundingbox is ALWAYS used as a constraint. This should probably be a configuration option but I am out of time for the day and can do this the next time I get time on the project.
Issue #GDP-453 - Add bbox of selected feature/s to CSW search to narrow returned data sets.

git-svn-id: https://cida-svn.er.usgs.gov/repos/dev/usgs/gdp/trunk@19924 622dc1fb-2224-0410-ab72-dc85a3975e1a