Added an on-hover text block that says "Select A Dataset" when a link is hovered over.
This is not done in the same tooltip style we have throughout the application. The reason being is that the rest of the tooltips in the application happen on-click and in order to get these specific <a> tags to activate on hover, we would have to rewrite the tooltip script a bit which would take an hour or so. 

Decided to hold off for now unless it becomes a requirement later
Issue #GDP-491 - Add tooltip to dataset selection

git-svn-id: https://cida-svn.er.usgs.gov/repos/dev/usgs/gdp/trunk@24646 622dc1fb-2224-0410-ab72-dc85a3975e1a