Tested against all datasets founds @ https://cida-svn.er.usgs.gov/repos/dev/usgs/gdp/trunk/gdp-process-wps/src/test/resources/wcs/
The main difference was that the GeoServer WCS does not have wcs as its default namespace and ArcGIS does. So we have to add the namespace to the search query when we're dealing with geoserver. 

Tested in Firefox, Chrome, IE7, IE8
Issue #GDP-480 - Handle all known WCS implementations at datset selection step

git-svn-id: https://cida-svn.er.usgs.gov/repos/dev/usgs/gdp/trunk@24793 622dc1fb-2224-0410-ab72-dc85a3975e1a